refactor(web-app): refactor dropdown menu #121
close #121
