+++ apps/web-app/src/components/DropDownMenu/components/index.ts
@@ -1,0 +1,3 @@
+export * from './DropDownItem/DropDownItem'
+export * from './DropDownListOfItems/DropDownListOfItems'
+
